Initial structure15

+++ tasks/mpi/sedova_o_max_of_vector_elements/include/ops_mpi.hpp
@@ -43,8 +43,8 @@ class TestMPITaskParallel : public ppc::core::Task {
   boost::mpi::communicator world;
 };
 
-std::vector<int> generate_random_vector(int size, int value);
-std::vector<std::vector<int>> generate_random_matrix(int rows, int cols, int value);
+std::vector<int> generate_random_vector(size_t size, size_t value);
+std::vector<std::vector<int>> generate_random_matrix(size_t rows, size_t cols, size_t value);
 int find_max_of_matrix(const std::vector<int> matrix);
 
 }  // namespace sedova_o_max_of_vector_elements_mpi
