+++ tasks/seq/sedova_o_max_of_vector_elements/include/ops_seq.hpp
@@ -6,8 +6,8 @@
 #include "core/task/include/task.hpp"
 
 namespace sedova_o_max_of_vector_elements_seq {
-std::vector<std::vector<int>> generate_random_matrix(int rows, int cols, int value);
-std::vector<int> generate_random_vector(int size, int value);
+std::vector<std::vector<int>> generate_random_matrix(size_t rows, size_t cols, size_t value);
+std::vector<int> generate_random_vector(size_t size, size_t value);
 int find_max_of_matrix(std::vector<int> matrix);
 
 class TestTaskSequential : public ppc::core::Task {
